+++ react/Mutations.js
@@ -1,0 +1,12 @@
+
+import addToCart from './mutations/addToCart.gql'
+import updateItems from './mutations/updateItems.gql'
+import updateOrderFormProfile from './mutations/updateOrderFormProfile.gql'
+import updateOrderFormShipping from './mutations/updateOrderFormShipping.gql'
+
+export default {
+  addToCart,
+  updateItems,
+  updateOrderFormProfile,
+  updateOrderFormShipping,
+}